Pulsar

+++ kafka-connect-pulsar/src/main/scala/com/datamountaineer/streamreactor/connect/pulsar/source/PulsarManager.scala
@@ -125,6 +125,6 @@ case class PulsarMessageConverter(convertersMap: Map[String, Converter],
   }
 
   def getRecords(target: util.ArrayList[SourceRecord]): Unit = {
-    QueueHelpers.drainWithTimeoutNoGauva(target, batchSize, pollingTimeout * 1000000, queue)
+    QueueHelpers.drainWithTimeoutNoGauva(target, batchSize, pollingTimeout.toLong * 1000000, queue)
   }
 }
